chore: use userInterestsConnection to fetch collected artists
